AddingSOLID principles.
Though we could put the display at the main(), I felt that it could be better if we store this inside a data structure. I still chose ArrayList because of its resizing , while retaining array capabilities. Was able to write and rewrite an excel. HDB officer and applicant have acess to the collect and show methods, while manager has access to editing. Add and delete functions have been added. Unfortunately, an area where this library falls short is at the data type. At present, library gives errors when parasing date and int, so I had to resort to converting everything to string. Will find an alternative later.